backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plansync-backend"
version = "0.1.0"
description = "PlanSync FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "cryptography>=44.0.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.7.0",
  "pwdlib[bcrypt]>=0.2.1",
  "psycopg[binary]>=3.2.0",
  "pyjwt>=2.10.0",
  "python-dateutil>=2.9.0",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.8.0"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
